Move all experimental kernel tests into kernel_tests
