--- numba_dpex/tests/experimental/kernel_api_overloads/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2023 - 2024 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
