+++ numba_dpex/tests/kernel_tests/test_async_kernel.py
