+++ numba_dpex/tests/kernel_tests/test_atomic_fence.py
