+++ numba_dpex/tests/kernel_tests/test_atomic_load_store_cmp_exchg.py
