+++ numba_dpex/tests/kernel_tests/test_barriers.py
