+++ numba_dpex/tests/kernel_tests/test_compiler_warnings.py
