+++ numba_dpex/tests/kernel_tests/test_memory_enum_compilation.py
